+++ src/main/java/com/boardgo/config/WebConfig.java
@@ -7,7 +7,7 @@ import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import com.boardgo.config.filter.LoggingFilter;
-import com.boardgo.config.interceptor.LogInterceptor;
+import com.boardgo.config.interceptor.LoggingInterceptor;
 
 import lombok.RequiredArgsConstructor;
 
@@ -15,7 +15,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
 
-	private final LogInterceptor interceptor;
+	private final LoggingInterceptor interceptor;
 
 	private static final String[] INTERCEPTOR_WHITE_LIST = {
 		"/css/**", "/js/**", "/images/**", "/fonts/**", "/*.html", "/v3/api-docs"
